+++ synapse-core/src/test/java/de/otto/synapse/messagestore/NonCompactingMessageStoreTest.java
@@ -40,6 +40,7 @@ public class NonCompactingMessageStoreTest {
     @Parameter
     public Supplier<WritableMessageStore> messageStoreBuilder;
 
+    @SuppressWarnings("Duplicates")
     @Test
     public void shouldCalculateUncompactedChannelPositions() {
         final WritableMessageStore messageStore = messageStoreBuilder.get();